moved 'create_file_func' to the utl folder so the functions can be used for both the front and back end
