+++ utl/create_file_func/client_queries.js
