+++ utl/create_file_func/create_client_root_index.js
