+++ utl/create_file_func/create_component_style.js
