+++ utl/create_file_func/mysql_scripts.js
