+++ utl/create_file_func/postgresql_scripts.js
